+++ src/itkUnaryCloudMaskImageFilter.h
@@ -8,43 +8,43 @@ namespace itk
   namespace Functor
   {
     // Description class
-    
+
     template<typename TInput, typename TOutput> class CloudMask
     {
     public:
       typedef typename NumericTraits< TInput >::AccumulateType AccumulatorType;
       CloudMask() {}
       ~CloudMask(){}
-      
+
       void SetCloudMask(const unsigned int val)
       {
-	m_cloud_mask_value = val;
+        m_cloud_mask_value = val;
       }
-      
+
       inline TOutput operator()(const TInput& B) const
       {
-	std::bitset<8> bits(B);
-	std::bitset<8> mask_bits(m_cloud_mask_value);
-	std::bitset<8> result;
-	result = bits & mask_bits;
-	if(result != 0)
-	  return static_cast<TOutput>(1);
-	else
-	  return static_cast<TOutput>(0);
-	
-	/* if(result.to_ulong()>0) */
-	/*   return static_cast<TOutput>(1); */
-	/* else */
-	/*   return static_cast<TOutput>(0); */
+        std::bitset<8> bits(B);
+        std::bitset<8> mask_bits(m_cloud_mask_value);
+        std::bitset<8> result;
+        result = bits & mask_bits;
+        if(result != 0)
+          return static_cast<TOutput>(1);
+        else
+          return static_cast<TOutput>(0);
+
+        /* if(result.to_ulong()>0) */
+        /*   return static_cast<TOutput>(1); */
+        /* else */
+        /*   return static_cast<TOutput>(0); */
       }
       bool operator==(const CloudMask &) const
       {
-	return true;
+        return true;
       }
-      
+
       bool operator!=(const CloudMask &) const
       {
-	return false;
+        return false;
       }
     private:
       int m_cloud_mask_value;
@@ -52,8 +52,8 @@ namespace itk
   }
   // Description functor
   template<typename TInputImage, typename TOutputImage> class UnaryCloudMaskImageFilter:
-    public 
-    UnaryFunctorImageFilter< TInputImage, TOutputImage, 
+    public
+    UnaryFunctorImageFilter< TInputImage, TOutputImage,
                             Functor::CloudMask< typename TInputImage::PixelType, typename TInputImage::PixelType > >
   {
   public:
@@ -65,7 +65,7 @@ namespace itk
 
   typedef SmartPointer< Self >       Pointer;
   typedef SmartPointer< const Self > ConstPointer;
-  
+
   void SetCloudMask(const unsigned int val)
   {
     this->GetFunctor().SetCloudMask(val);